requirements.txt
pandas
tqdm
scikit-learn
openai
python-dotenv
torch-2.8.0
transformers
bitsandbytes
google-generativeai
nest-asyncio